Gate `ModuleTranslation` and its methods on `compile`
